fix: BigDecimal으로 일관성 있게 고침

+++ src/main/java/com/part3/deokhugam/dto/book/PopularBookDto.java
@@ -14,7 +14,7 @@ public record PopularBookDto(
         String thumbnailUrl,
         Period period,
         Integer rank,
-        Double score,
+        BigDecimal score,
         Integer reviewCount,
         BigDecimal rating,
         Instant createdAt
